front-end log: - change discount form to additional discount. - Add percent input format for discount. back-end log: - change add order validation. - change order details schema.

+++ database/migrations/2022_12_02_110217_create_order_details_table.php
@@ -21,7 +21,7 @@ return new class extends Migration
             $table->json('clothes');
             $table->integer('service_quantity');
             $table->bigInteger('subtotal');
-            $table->float('discount', 4, 3);
+            $table->bigInteger('discount');
             $table->bigInteger('total');
             $table->timestamps();
         });
